Create countTwo.R
Similar to countWords.R
Takes a list of phrases or sentences and returns a dataframe listing each pair of words found and the number of times they appeared together.
Finds the rate of occurrence of pairs of words. Later version will allow this function to find the rate of occurrence of sequences n words long.